+++ semana19/arquitetura/src/data/deleteUserQueries.ts
